requirements.txt
fastapi
uvicorn
numpy
scikit-learn
joblib
pandas
gspread
google-auth